pyproject.toml
[tool.poetry]
name = "gitlab-api"
version = "0.1.0"
description = "To interact with GitLab API"
authors = ["Daniel Montenegro <daniel.montenegro@live-eo.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.32.3"
python-gitlab = "^5.6.0"
flask = "^3.1.0"
plotly = "^6.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
